requirements.txt
pypdf>=4.0.0
fastapi>=0.115.0
uvicorn>=0.30.0
PyYAML>=6.0.0
